requirements.txt
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
google-cloud-storage==1.42.3
requests==2.26.0